# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trade/buy:
    post:
      operationId: buyShares
      tags: [trading]
      summary: Buy SPCX shares using available cash balance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BuyInput"
      responses:
        "200":
          description: Updated holdings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holdings"
        "400":
          description: Invalid input or insufficient cash balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Investor not approved for trading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trade/sell:
    post:
      operationId: sellShares
      tags: [trading]
      summary: Sell SPCX shares (only permitted while the admin has selling enabled)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellInput"
      responses:
        "200":
          description: Updated holdings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holdings"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Selling disabled or investor not approved for trading
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Investor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Holdings:
      type: object
      properties:
        investorId:
          type: integer
        shares:
          type: string
        avgCost:
          type: string
        cashBalance:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - investorId
        - shares
        - avgCost
        - cashBalance
        - updatedAt
...
    BuyInput:
      type: object
      properties:
        email:
          type: string
        amountUsd:
          type: number
          minimum: 0.01
      required:
        - email
        - amountUsd
    SellInput:
      type: object
      properties:
        email:
          type: string
        shares:
          type: number
          minimum: 0.0001
      required:
        - email
        - shares
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error