# lib/api-zod/src/generated/types/buyInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BuyInput {
  email: string;
  /** @minimum 0.01 */
  amountUsd: number;
}