# lib/api-zod/src/generated/types/sellInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SellInput {
  email: string;
  /** @minimum 0.0001 */
  shares: number;
}